added region-movement

+++ gameServer/src/region/controller.cpp
@@ -4,23 +4,23 @@ region::controller::controller() {
     
 }
 
-void region::controller::open_region(const uint64_t id) {
+void region::controller::open_region(const uint32_t id) {
     std::lock_guard<std::mutex> lg(m_mutex);
     m_regionMap.emplace(id, id);
     m_regionUpdateQueue.insert(id);
 }
 
-void region::controller::close_region(const uint64_t id) {
+void region::controller::close_region(const uint32_t id) {
     std::lock_guard<std::mutex> lg(m_mutex);
     m_regionMap.erase(id);
 }
 
-bool region::controller::is_open(const uint64_t id) {
+bool region::controller::is_open(const uint32_t id) {
     std::lock_guard<std::mutex> lg(m_mutex);
     return m_regionMap.count(id) != 0;
 }
 
-util::locked_ref<region::context> region::controller::get_region(const uint64_t id) {
+util::locked_ref<region::context> region::controller::get_region(const uint32_t id) {
     std::lock_guard<std::mutex> lg(m_mutex);
     
     if (m_regionMap.count(id) == 0) {
@@ -33,7 +33,7 @@ util::locked_ref<region::context> region::controller::get_region(const uint64_t 
 }
 
 void region::controller::update() {
-    uint64_t current;
+    uint32_t current;
     if (m_regionUpdateQueue.wait_and_get(&current)) {
         update_queue::timed_obj tObj(current);
         auto region = get_region(current);
